requirements.txt
# Core libraries
tensorflow>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pymongo>=4.5.0
python-dotenv>=1.0.0

# Data processing and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
geopandas>=0.13.0

# Jupyter and development tools
jupyter
notebook
jupyterlab
ipykernel
ipython
ipywidgets

# Web and API
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Serialization and storage
joblib>=1.3.0
pickle5>=0.0.12; python_version < '3.8'

# Documentation
sphinx>=7.1.0